requirements.txt
arrow==1.4.0
asgiref==3.11.1
attrs==25.4.0
binaryornot==0.4.4
boto3==1.42.25
botocore==1.42.25
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cookiecutter==2.6.0
cssbeautifier==1.15.4
Django==6.0.3
django-allauth==65.13.1
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
djlint==1.36.4
dotenv==0.9.9
drf-spectacular==0.29.0
EditorConfig==0.17.1
ezdxf==1.4.3
fonttools==4.62.1
geopandas==1.1.2
gunicorn==23.0.0
honcho==2.0.0
idna==3.11
inflection==0.5.1
Jinja2==3.1.6
jmespath==1.0.1
jsbeautifier==1.15.4
json5==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mbutil==0.3.0
mdurl==0.1.2
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pathspec==1.0.3
pillow==12.1.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.12.1
pyogrio==0.12.1
pyparsing==3.3.2
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
s3transfer==0.16.0
shapely==2.1.2
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.11.0
